app/requirements.txt
qdrant-client
sentence-transformers
groq
streamlit
python-dotenv
langchain-text-splitters
pandas
tqdm
rank-bm25
